add serializers consumer, consumer with password to core
